# hugolib: Enable nested sections