Fixes #465 

See http://hugotest.bep.is/ for a demo site.