@bep 2017-06-04
This is mostly done now, but a fair amount of TODOs in there that need to be handled.

Performance should be just about the same, looks slightlly faster:

```
▶ ./benchSite.sh "TOML,num_root_sections=5,num_pages=1000" 

github.com/spf13/hugo  master ✔                                                                                                                                                        19h30m
▶ benchcmp pre.bench post.bench
benchmark                                                           old ns/op     new ns/op     delta
BenchmarkSiteBuilding/TOML,num_root_sections=5,num_pages=1000-4     114543797     111523225     -2.64%

benchmark                                                           old allocs     new allocs     delta
BenchmarkSiteBuilding/TOML,num_root_sections=5,num_pages=1000-4     595315         602237         +1.16%

benchmark                                                           old bytes     new bytes     delta
BenchmarkSiteBuilding/TOML,num_root_sections=5,num_pages=1000-4     68144409      68539393      +0.58%
```